[FIR] Remove strange code from `getNarrowedDownSymbol()`
This change is important for IDE/our LL tests. This code is incorrect
until a `lazyResolveToPhase(TYPES)` is inserted before acessing the type
of the field, but knowing a property's type doesn't require knowing it
at all. If there's an explicit field, there's no regular initializer,
meaning the type is already written explicitly. And it's not affected
by the field data (otherwise, we would have to wait until
`IMPLICIT_TYPES_BODY_RESOLVE`).

^KT-68082