[IR] Fix validating unbound symbols but not tree consistency
This fixes the checkUnboundSymbols() function, which specifies
checkUnboundSymbols = true, checkTreeConsistency = false.
Because only checkTreeConsistency was tested in IrValidator, it was a
no-op.

Also, relax the rule for early return in the case of duplicated nodes.
There is, unfortunately, a case where IR nodes are duplicated, but not
in a way that causes an infinite loop (KT-79807). But in
checkUnboundSymbols() we don't yet want to fail on that, as we only
check symbols.

KT-77819
KT-79807